python-scraper/requirements.txt
httpx>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
click>=8.1.0
schedule>=1.2.0
tenacity>=8.2.0
beautifulsoup4>=4.12.0

# Dev / Test dependencies
pytest>=8.0.0
pytest-httpx>=0.30.0
respx>=0.21.0
